requirements.txt
fastapi
uvicorn[standard]
torch==2.2.2+cpu
torchvision==0.17.2+cpu
pillow
vit-pytorch
python-multipart
